+++ src/components/button.js
@@ -4,6 +4,7 @@ import Color from 'color';
 import Text from 'components/text';
 import { colors } from 'styles/color';
 import noop from 'util/noop';
+import withCSS from 'util/with-css';
 
 const COLOR_INTENSITY_RATIO = 0.08;
 
@@ -165,4 +166,11 @@ class Button extends Component {
   }
 }
 
-export default Button;
+export default withCSS({
+  key: 'button',
+  css: `
+    button::-moz-focus-inner {
+      border: 0;
+    }
+  `,
+})(Button);