+++ CHANGELOG.md
@@ -10,11 +10,13 @@
 
 * `Button` no longer automatically blurs on `mouseleave` events
 * Unnecessary explicit `Button` prop `onClick` removed
+* `Button` applies click style for touch events
 * `Button` now has a dedicated focus style, identical to the existing hover style
 * Minor tweaks to the hover and active appearance of `Button`
 * Fixed regression introduced in `Modal` where it would not expand to full height when the viewport height is too small
 * `Checkbox` no longer uses a pointer hand when it is `disabled`
 * `Link` has dedicated focus and active style when using keyboard for navigation
+* `Link` applies click style for touch events
 
 ### Removed